Revert "Whoops..."
This reverts commit 4c8c265490473d490c32b559ac6aa4414a38ef46.

+++ code/breather.c
@@ -197,7 +197,7 @@ int main() {
 	
 	// Time how long the operation took
 	gettimeofday(&end, NULL);
-	double delta = ((end.tv_sec  - start.tv_sec) * 1000000u + 
+	long double delta = ((end.tv_sec  - start.tv_sec) * 1000000u + 
 	end.tv_usec - start.tv_usec) / 1.e6;
 	printf("Total time=%f seconds\n", delta);
 }
